requirements.txt
PyYAML
transformers
datasets
peft
torch
accelerate
numpy
wandb
